issue #1: rename factory app_creator
